fixed deletion bug.

+++ ManufacturersService/src/main/java/com/grinch/ManufacturersService/Services/ManufacturersService.java
@@ -1,11 +1,10 @@
 package com.grinch.ManufacturersService.Services;
 
 import java.util.Optional;
-
+import javax.jms.Topic;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-
+import org.springframework.transaction.annotation.Transactional;
 import com.grinch.ManufacturersService.BusinessLogic.Entites.Manufacturer;
 import com.grinch.ManufacturersService.Exceptions.ResourceAlreadyExistsException;
 import com.grinch.ManufacturersService.Exceptions.ResourceNotFoundException;
@@ -14,20 +13,21 @@ import com.grinch.ManufacturersService.Repositories.ManufacturersRepository;
 
 
 @Service
+@Transactional
 public class ManufacturersService {
 	@Autowired
 	private ManufacturersRepository repository;
 	@Autowired
 	private ManufacturerPublisher publisher;
-	@Value("${AcitveMQ.Topics.manufacturersTopic}")
-	private String manufacturersTopic;
+	@Autowired
+	private Topic manuTopic;
 	
 	public Manufacturer getManufacturer(Long id) throws Exception {
-		Optional<Manufacturer> Manufacturer = repository.findById(id);
-		if(Manufacturer.isEmpty()) {
+		Optional<Manufacturer> manufacturer = repository.findById(id);
+		if(manufacturer.isEmpty()) {
 			throw new ResourceNotFoundException("Manufacturer with id " + id + " was not found.");
 		}
-		return Manufacturer.get();
+		return manufacturer.get();
 	}
 	
 	public Manufacturer postManufacturer(Manufacturer manufacturer) throws Exception {
@@ -36,7 +36,6 @@ public class ManufacturersService {
 			throw new ResourceAlreadyExistsException("Manufacturer with name " + manufacturer.getName() + " is already exists.");
 		}
 		Manufacturer result = repository.save(manufacturer);
-		publisher.publishManufacturer(manufacturersTopic, manufacturer);
 		return result;
 	}
 	
@@ -52,7 +51,12 @@ public class ManufacturersService {
 		return repository.save(manufacturer);
 	}
 	
-	public void deleteManufacturer(Long id) {
+	public void deleteManufacturer(Long id) throws ResourceNotFoundException {
+		Optional<Manufacturer> manufacturer = repository.findById(id);
+		if(manufacturer.isEmpty()) {
+			throw new ResourceNotFoundException("Manufacturer with id " + id + " was not found.");
+		}
 		repository.deleteById(id);
+		publisher.publishManufacturer(manuTopic, manufacturer.get()); //publish manufacturer in topic to delete related items.
 	}
 }
