+++ ReceiversService/src/main/java/com/grinch/ReceiversService/Services/ReceiversService.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.grinch.ReceiversService.BusinessLogic.Entities.Receiver;
 import com.grinch.ReceiversService.Exceptions.ResourceAlreadyExistsException;
 import com.grinch.ReceiversService.Exceptions.ResourceNotFoundException;
@@ -26,7 +25,7 @@ public class ReceiversService {
 	
 	public Receiver postReceiver(Receiver Receiver) throws ResourceAlreadyExistsException {
 		if(!repository.findByNameAndManufacturerReference_Id(Receiver.getName(), Receiver.getManufacturerReference().getId()).isEmpty()) {
-			throw new ResourceAlreadyExistsException("Receiver with name " + Receiver.getName() + " is already exists for this manufacturer.");
+			throw new ResourceAlreadyExistsException("Receiver with name " + Receiver.getName() + " is already exists for this Receiver.");
 		}
 		return repository.save(Receiver);
 	}
@@ -37,12 +36,15 @@ public class ReceiversService {
 		}
 		Optional<Receiver> test = repository.findByNameAndManufacturerReference_Id(Receiver.getName(), Receiver.getManufacturerReference().getId());
 		if((!test.isEmpty())&&test.get().getId()!=Receiver.getId()) {
-			throw new ResourceAlreadyExistsException("Receiver with name " + Receiver.getName() + " is already exists for this manufacturer.");
+			throw new ResourceAlreadyExistsException("Receiver with name " + Receiver.getName() + " is already exists for this Receiver.");
 		}
 		return repository.save(Receiver);
 	}
 	
-	public void deleteReceiver(Long id) {
+	public void deleteReceiver(Long id) throws ResourceNotFoundException {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Receiver with id " + id + " was not found.");
+		}
 		repository.deleteById(id);
 	}
 }
